+++ src/main/java/com/example/model/Country.java
@@ -1,20 +1,22 @@
 package com.example.model;
 
 public class Country {
-    private Long id;
+    private Integer id;
     private String countryName;
     private String imgUrl;
 
     public Country() {
     }
-    public Country(Long id, String countryName) {
+
+    public Country(Integer id, String countryName) {
         this.id = id;
         this.countryName = countryName;
     }
-    public Long getId() {
+    
+    public Integer getId() {
         return id;
     }
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
     public String getCountryName() {
